+++ src/XPike.Metrics.DataDog/IDependencyCollectionExtensions.cs
@@ -6,5 +6,11 @@ namespace XPike.Metrics.DataDog
     {
         public static IDependencyCollection AddXPikeDataDogMetrics(this IDependencyCollection collection) =>
             collection.LoadPackage(new XPike.Metrics.DataDog.Package());
+
+        public static IDependencyCollection UseXPikeDataDogMetrics(this IDependencyCollection collection)
+        {
+            collection.ResetCollection<IMetricsProvider>();
+            return collection.AddXPikeDataDogMetrics();
+        }
     }
 }