+++ src/XPike.Metrics.Microsoft/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace XPike.Metrics.Microsoft
-{
-    public class Class1
-    {
-    }
-}