+++ src/XPike.Metrics/Console/ConsoleMetricsProvider.cs
@@ -9,8 +9,19 @@ namespace XPike.Metrics.Console
     {
         private readonly ConcurrentQueue<Action> _events = new ConcurrentQueue<Action>();
 
-        public void Send<T>(MetricType metric, string name, T value, double sampleRate, IEnumerable<string> tags) =>
+        public void Send<T>(MetricType metric, string name, T value, double sampleRate, IEnumerable<string> tags)
+        {
+            var fg = System.Console.ForegroundColor;
+            var bg = System.Console.BackgroundColor;
+
+            System.Console.BackgroundColor = ConsoleColor.Black;
+            System.Console.ForegroundColor = ConsoleColor.DarkCyan;
+
             System.Console.WriteLine($"[{metric}] {name}: {value} (sampleRate={sampleRate}, tags={string.Join(",", tags)})");
+
+            System.Console.ForegroundColor = fg;
+            System.Console.BackgroundColor = bg;
+        }
 
         public void Add<T>(MetricType metric, string name, T value, double sampleRate, IEnumerable<string> tags) =>
             _events.Enqueue(() => Send<T>(metric, name, value, sampleRate, tags));