+++ src/XPike.Metrics/IDependencyCollectionExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using XPike.IoC;
+using XPike.Metrics.Console;
 
 namespace XPike.Metrics
 {
@@ -6,5 +7,20 @@ namespace XPike.Metrics
     {
         public static IDependencyCollection AddXPikeMetrics(this IDependencyCollection collection) =>
             collection.LoadPackage(new XPike.Metrics.Package());
+
+        public static IDependencyCollection AddXPikeConsoleMetrics(this IDependencyCollection collection)
+        {
+            collection.AddXPikeMetrics()
+                .AddSingletonToCollection<IMetricsProvider, IConsoleMetricsProvider>(container =>
+                container.ResolveDependency<IConsoleMetricsProvider>());
+
+            return collection;
+        }
+
+        public static IDependencyCollection UseXPikeConsoleMetrics(this IDependencyCollection collection)
+        {
+            collection.ResetCollection<IMetricsProvider>();
+            return collection.AddXPikeConsoleMetrics();
+        }
     }
 }